more refactor of names
