+++ api/api.cbp
@@ -60,7 +60,7 @@
 		<Unit filename="pt_gs_k.h" />
 		<Unit filename="pt_hs_k.h" />
 		<Unit filename="pt_ss_k.h" />
-		<Unit filename="timeseries.cpp" />
+		<Unit filename="time_series.cpp" />
 		<Unit filename="timeseries.h" />
 		<Extensions>
 			<code_completion />
