+++ api/api.vcxproj
@@ -13,7 +13,7 @@
   <ItemGroup>
     <ClCompile Include="api.cpp" />
     <ClCompile Include="api_serialization.cpp" />
-    <ClCompile Include="timeseries.cpp" />
+    <ClCompile Include="time_series.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="api.h" />
@@ -21,7 +21,7 @@
     <ClInclude Include="hbv_stack.h" />
     <ClInclude Include="pt_gs_k.h" />
     <ClInclude Include="pt_ss_k.h" />
-    <ClInclude Include="timeseries.h" />
+    <ClInclude Include="time_series.h" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
     <ProjectGuid>{FDFDAA14-75D7-455C-AAE4-E89A4E43C062}</ProjectGuid>
