+++ api/time_series.cpp
@@ -1,5 +1,5 @@
 #include "core/core_pch.h"
-#include "timeseries.h"
+#include "time_series.h"
 #include <dlib/statistics.h>
 
 namespace shyft{
