+++ core/core.cbp
@@ -5,7 +5,7 @@
 		<Option title="core" />
 		<Option pch_mode="2" />
 		<Option compiler="gcc" />
-		<Option virtualFolders="methods/;interpolation/;method_stacks/;time_and_calendar/;geo/;time_series/;compiler_etc/;experimental/;optimizers/;serialization/;" />
+		<Option virtualFolders="methods/;interpolation/;method_stacks/;time_and_calendar/;geo/;time_series.h/;compiler_etc/;experimental/;optimizers/;serialization/;" />
 		<Option show_notes="0">
 			<notes>
 				<![CDATA[The core parts, - basically our target is a header-only concept, but there is still a need for a tiny static library]]>
@@ -175,10 +175,10 @@
 			<Option virtualFolder="methods/" />
 		</Unit>
 		<Unit filename="time_axis.h">
-			<Option virtualFolder="time_series/" />
+			<Option virtualFolder="time_series.h/" />
 		</Unit>
-		<Unit filename="timeseries.h">
-			<Option virtualFolder="time_series/" />
+		<Unit filename="time_series.h">
+			<Option virtualFolder="time_series.h/" />
 		</Unit>
 		<Unit filename="unit_conversion.h">
 			<Option virtualFolder="compiler_etc/" />
