+++ core/core.vcxproj
@@ -145,7 +145,7 @@
     <ClInclude Include="routing.h" />
     <ClInclude Include="sceua_optimizer.h" />
     <ClInclude Include="skaugen.h" />
-    <ClInclude Include="timeseries.h" />
+    <ClInclude Include="time_series.h" />
     <ClInclude Include="region_model.h" />
     <ClInclude Include="time_axis.h" />
     <ClInclude Include="unit_conversion.h" />
