+++ test/kalman_test.cpp
@@ -1,7 +1,7 @@
 #include "test_pch.h"
 #include "mocks.h"
 #include "api/api.h" // looking for GeoPointSource, and TemperatureSource(realistic case)
-#include "api/timeseries.h" // looking for apoint_ts, the api exposed ts-type(realistic case)
+#include "api/time_series.h" // looking for apoint_ts, the api exposed ts-type(realistic case)
 #include "core/kalman.h"
