+++ test/kriging_test.cpp
@@ -1,6 +1,6 @@
 #include "test_pch.h"
 #include "core/kriging.h"
-#include "core/timeseries.h"
+#include "core/time_series.h"
 #include "core/geo_point.h"
