+++ test/pt_gs_k_test.cpp
@@ -5,7 +5,7 @@
 #include "core/geo_cell_data.h"
 #include "core/geo_point.h"
 #include "mocks.h"
-#include "core/timeseries.h"
+#include "core/time_series.h"
 #include "core/utctime_utilities.h"
 
 // Some typedefs for clarity
