+++ test/region_model_test.cpp
@@ -6,7 +6,7 @@
 #include "core/geo_cell_data.h"
 
 
-#include "core/timeseries.h"
+#include "core/time_series.h"
 #include "core/inverse_distance.h"
 #include "core/bayesian_kriging.h"
