+++ test/serialization_test.cpp
@@ -2,7 +2,7 @@
 
 #include "core/utctime_utilities.h"
 #include "core/time_axis.h"
-#include "api/timeseries.h"
+#include "api/time_series.h"
 
 #include <boost/archive/binary_iarchive.hpp>
 #include <boost/archive/binary_oarchive.hpp>
